Fix compilation errors and most of the tests
