+++ GridKit/Model/PhasorDynamics/Load/Load.hpp
@@ -36,6 +36,7 @@ namespace GridKit
       using Component<scalar_type, index_type>::alpha_;
       using Component<scalar_type, index_type>::y_;
       using Component<scalar_type, index_type>::yp_;
+      using Component<scalar_type, index_type>::abs_tol_;
       using Component<scalar_type, index_type>::tag_;
       using Component<scalar_type, index_type>::wb_;
       using Component<scalar_type, index_type>::h_;
