+++ GridKit/Model/PhasorDynamics/LoadZIP/LoadZIP.hpp
@@ -36,6 +36,7 @@ namespace GridKit
       using Component<scalar_type, index_type>::alpha_;
       using Component<scalar_type, index_type>::y_;
       using Component<scalar_type, index_type>::yp_;
+      using Component<scalar_type, index_type>::abs_tol_;
       using Component<scalar_type, index_type>::tag_;
       using Component<scalar_type, index_type>::wb_;
       using Component<scalar_type, index_type>::h_;
@@ -64,6 +65,7 @@ namespace GridKit
       int allocate() override final;
       int initialize() override final;
       int tagDifferentiable() override final;
+      int setAbsoluteTolerance(RealT rel_tol) override final;
       int evaluateResidual() override final;
       int evaluateJacobian() override final;
