+++ GridKit/Model/PhasorDynamics/Stabilizer/IEEEST/Ieeest.hpp
@@ -66,6 +66,7 @@ namespace GridKit
         using Component<scalar_type, index_type>::nnz_;
         using Component<scalar_type, index_type>::size_;
         using Component<scalar_type, index_type>::tag_;
+        using Component<scalar_type, index_type>::abs_tol_;
         using Component<scalar_type, index_type>::time_;
         using Component<scalar_type, index_type>::y_;
         using Component<scalar_type, index_type>::yp_;
@@ -95,6 +96,7 @@ namespace GridKit
         int verify() const override final;
         int initialize() override final;
         int tagDifferentiable() override final;
+        int setAbsoluteTolerance(RealT rel_tol) override final;
         int evaluateResidual() override final;
         int evaluateJacobian() override final;
