+++ GridKit/Model/PhasorDynamics/SynchronousMachine/GENSALwS/Gensal.hpp
@@ -70,6 +70,7 @@ namespace GridKit
       using Component<scalar_type, index_type>::nnz_;
       using Component<scalar_type, index_type>::size_;
       using Component<scalar_type, index_type>::tag_;
+      using Component<scalar_type, index_type>::abs_tol_;
       using Component<scalar_type, index_type>::time_;
       using Component<scalar_type, index_type>::y_;
       using Component<scalar_type, index_type>::yp_;
@@ -100,6 +101,7 @@ namespace GridKit
       int verify() const override final;
       int initialize() override final;
       int tagDifferentiable() override final;
+      int setAbsoluteTolerance(RealT rel_tol) override final;
       int evaluateResidual() override final;
 
       // Still to be implemented
