+++ GridKit/Model/PowerFlow/Generator4Governor/Generator4Governor.cpp
@@ -2,6 +2,7 @@
 #define _USE_MATH_DEFINES
 #include "Generator4Governor.hpp"
 
+#include <algorithm>
 #include <cmath>
 #include <iostream>
 
@@ -165,6 +166,13 @@ namespace GridKit
     tag_[static_cast<size_t>(offsetGov_ + 1)] = true;
     tag_[static_cast<size_t>(offsetGov_ + 2)] = false;
 
+    return 0;
+  }
+
+  template <class ScalarT, typename IdxT>
+  int Generator4Governor<ScalarT, IdxT>::setAbsoluteTolerance(RealT rel_tol)
+  {
+    std::fill(abs_tol_.begin(), abs_tol_.end(), rel_tol);
     return 0;
   }
