+++ GridKit/Model/PowerFlow/Generator4Governor/Generator4Governor.hpp
@@ -47,6 +47,7 @@ namespace GridKit
     int allocate();
     int initialize();
     int tagDifferentiable();
+    int setAbsoluteTolerance(RealT);
     int evaluateResidual();
     int evaluateJacobian();
     int evaluateIntegrand();
