+++ GridKit/Model/PowerFlow/Generator4Param/Generator4Param.hpp
@@ -46,6 +46,7 @@ namespace GridKit
     int allocate();
     int initialize();
     int tagDifferentiable();
+    int setAbsoluteTolerance(RealT);
     int evaluateResidual();
     int evaluateJacobian();
     int evaluateIntegrand();
