+++ GridKit/Solver/Dynamic/Ida.hpp
@@ -59,7 +59,7 @@ namespace AnalysisManager
       int configureLinearSolverDense();
       int getDefaultInitialCondition();
       int setIntegrationTime(RealT t_init, RealT t_final, int nout);
-      int initializeSimulation(RealT t0, bool findConsistent = false);
+      int initializeSimulation(RealT t0, bool findConsistent = true);
 
       int runSimulation(RealT tf, int nout = 1, std::optional<std::function<void(RealT)>> step_callback = {});
       int deleteSimulation();
