+++ examples/Experimental/GenInfiniteBus/GenInfiniteBus.cpp
@@ -55,8 +55,8 @@ int main()
   // create initial condition after a fault
   {
     idas.getSavedInitialCondition();
-    idas.initializeSimulation(t_init);
     gen.V() = 0.0;
+    idas.initializeSimulation(t_init);
     idas.runSimulation(t_clear, 20);
     gen.V() = 1.0;
     idas.saveInitialCondition();
