+++ examples/PhasorDynamics/Small/TenGen/Classical/TenGenClassical.cpp
@@ -170,13 +170,13 @@ int main()
 
   // Introduce fault to ground and run for 0.1s
   fault.setStatus(1);
-  ida.initializeSimulation(1.0, true);
+  ida.initializeSimulation(1.0);
   nout = static_cast<int>(std::round((1.1 - 1.0) / dt));
   ida.runSimulation(1.1, nout, output_cb);
 
   // Clear fault and run until t = 10s.
   fault.setStatus(0);
-  ida.initializeSimulation(1.1, true);
+  ida.initializeSimulation(1.1);
   nout = static_cast<int>(std::round((10.0 - 1.1) / dt));
   ida.runSimulation(10.0, nout, output_cb);
   real_type stop = static_cast<real_type>(clock());
