+++ examples/PhasorDynamics/Tiny/TwoBus/Ieeet1/TwoBusIeeet1Json.cpp
@@ -145,14 +145,14 @@ int main(int argc, const char* argv[])
   // Introduce fault and run for the next 0.1s
   // fault.setStatus(true);
   fault->setStatus(true);
-  ida.initializeSimulation(1.0, true);
+  ida.initializeSimulation(1.0);
   nout = static_cast<int>(std::round((1.1 - 1.0) / dt));
   ida.runSimulation(1.1, nout, output_cb);
 
   // Clear the fault and run until t = 10s.
   // fault.setStatus(false);
   fault->setStatus(false);
-  ida.initializeSimulation(1.1, true);
+  ida.initializeSimulation(1.1);
   nout = static_cast<int>(std::round((10.0 - 1.1) / dt));
   ida.runSimulation(10.0, nout, output_cb);
   real_type stop = static_cast<real_type>(clock());
